Initialize main game classes and project folders
Created main program folders and classes to improve organization of the project. Added Renderhandler class, item class, SnakeGame class, States class, and ItemExample class. 

Item class and ItemExample class were added to help other programmers better integrate their code with pygame. Item class defines how each item in the game should be defined so that there is a little variation in implementation. It also allows the RenderHandler class to easily render each item, based on the specifications defined in the item subclasses.

All objects can be rendered based on a 2d array representing the game in numbers. This was done so that the team can program using tools they are comfortable with rather than relying on pygame. However, if the team decides to switch to directly working with pygame, the item's class also has that functionality. 

The RenderHandler class has all fully working methods, however, it's update function currently uses temp objects and needs to be switched to using real game objects once they are implemented.

The SnakeGame class is not fully completed but works to ensure that the code runs and can be tested. The main game loop can control rendering and updating the game based on the 2d array. However, more work needs to be done with this class so that other game states work. Also, so that keyboard input can be detected.