Initialise Player class and implement move method
I set Player to inherit attributes and methods from the Item Class, making the colour of the snake green to match the traditional game. I also implemented the move method which moves the position of the snake in the game based on key presses (W, A, S, D or the arrows).