Main game board is now rendered and properly updated
The snake game class now has a game board parameter. The main loop is now able to read and edit that board to display and update the content of the game.

Attributes and code needs to be cleaned up for readability. Right now classes are very unorganised. Waiting for game ticks to be added to update function in snake game class. Also waiting for player class to be fixed.