Player now moves on screen
The random board generations have been removed. The player class has been heavily modified so that the player can move. However, most functionality has been removed and needs to be added back.

There are a lot of glitches in this version that need to be fixed. For instance the player can go past the left board. Also the game randomly crashes on start. Also we need to fix the player class so that each once food is consumed the snake grows.