Fix screen mapping and add boarder to map
Swap which dimension the board is initialized with. Original code set
the board as width and height, but this was not scaling well to the
pygame screen. As such the original game was smaller than the pygame map
causing the game to crash once the player went to the right hand of the
stage. Now the board is set to height and width which better fits the
screen.

Create "initialize_board" function in board.py, rather than starting the
board with all empty spots. The purpose of this function to create an
empty map and then add walls to the borders.

However, the game still has the following bugs - Player passes through
walls - Player does not grow properly - Player does does not interact
with food

Signed-off-by: jaffarys <syed.jaffary@mail.utoronto.ca>